% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geocodes.R
\docType{data}
\name{historical_geocode_pdfs}
\alias{historical_geocode_pdfs}
\title{URLs of legacy geocode PDFs from HUD}
\format{
An object of class \code{character} of length 3.
}
\usage{
historical_geocode_pdfs
}
\description{
URLs of legacy geocode PDFs from HUD for use with \code{hud_geocodes}
}
\keyword{datasets}
