% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud.extract-package.R
\docType{package}
\name{hud.extract-package}
\alias{hud.extract}
\alias{hud.extract-package}
\title{hud.extract: Extracts R-friendly Specifications from the HUD Specification PDF}
\description{
Create useful R objects for working with HUD Export data
}
\author{
\strong{Maintainer}: Stephen Holsenbeck \email{hmisdataanalyst@cohhio.org}

Authors:
\itemize{
  \item Stephen Holsenbeck \email{sholsen@alumni.emory.edu}
}

}
\keyword{internal}
