% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_value_extract.R
\name{hud_dimensions}
\alias{hud_dimensions}
\title{Set the dimensions of the main content area for the PDF}
\usage{
hud_dimensions(file = hud_spec_pdf, pg = 20)
}
\arguments{
\item{file}{\code{(character)} Path to file}

\item{pg}{\code{(numeric)} Page number to use. Pages with a table filling the entire content will provide the most accurate boundaries of the main content area.}
}
\value{
\code{(named list)} with X & Y minimums and maximums
}
\description{
Use \link[tabulizer]{locate_areas} to set the x & y minimums and maximums for the main content area excluding header and footer regions for future use.
}
