% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_spec_extract.R
\name{hud_export_specs}
\alias{hud_export_specs}
\title{Extract HUD Export Item Specifications}
\usage{
hud_export_specs(hud_pdf_data)
}
\arguments{
\item{hud_pdf_data}{\code{tibble} output from \code{hud_pdf_data}}
}
\value{
\code({list}) of tibbles with specifications from tables in the PDF
}
\description{
Extract the specifications for each of the HUD Export Items from the HUD Specifications PDF
}
