% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_spec_extract.R
\name{hud_pdf_data}
\alias{hud_pdf_data}
\title{Create a pdf data list from the HUD 2022 HMIS Specs}
\usage{
hud_pdf_data(
  pdf = "https://hudhdx.info/Resources/Vendors/HMIS_CSV_Specifications_FY2022_v1.0.pdf",
  font_info = TRUE,
  opw = "",
  upw = ""
)
}
\arguments{
\item{pdf}{file path or raw vector with pdf data}

\item{font_info}{if TRUE, extract font-data for each box. Be careful, this
requires a very recent version of poppler and will error otherwise.}

\item{opw}{string with owner password to open pdf}

\item{upw}{string with user password to open pdf}
}
\description{
Utilities based on libpoppler for extracting text, fonts, attachments
and metadata from a pdf file.
}
\details{
The \code{\link[pdftools]{pdf_text}} function renders all textboxes on a text canvas
and returns a character vector of equal length to the number of pages in the
PDF file. On the other hand, \code{\link[pdftools]{pdf_data}} is more low level and
returns one data frame per page, containing one row for each textbox in the PDF.

Note that \code{\link[pdftools]{pdf_data}} requires a recent version of libpoppler
which might not be available on all Linux systems.
When using \code{\link[pdftools]{pdf_data}} in R packages, condition use on
\code{poppler_config()$has_pdf_data} which shows if this function can be
used on the current system. For Ubuntu 16.04 (Xenial) and 18.04 (Bionic)
you can use \href{https://github.com/ropensci/pdftools#installation}{the PPA}
with backports of Poppler 0.74.0.

Poppler is pretty verbose when encountering minor errors in PDF files,
in especially \code{\link[pdftools]{pdf_text}}. These messages are usually safe
to ignore, use \code{\link{suppressMessages}} to hide them altogether.
}
\examples{
# Just a random pdf file
pdf_file <- file.path(R.home("doc"), "NEWS.pdf")
info <- pdf_info(pdf_file)
text <- pdf_text(pdf_file)
fonts <- pdf_fonts(pdf_file)
files <- pdf_attachments(pdf_file)
}
\seealso{
Other pdftools: 
\code{\link[pdftools]{pdf_ocr_text}()},
\code{\link[pdftools]{qpdf}},
\code{\link[pdftools]{rendering}}
}
