% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_spec_extract.R
\docType{data}
\name{hud_spec_2020}
\alias{hud_spec_2020}
\title{2021 HUD Specifications PDF}
\format{
An object of class \code{character} of length 1.
}
\usage{
hud_spec_2020
}
\description{
2021 HUD Specifications PDF
}
\keyword{datasets}
