% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_spec_extract.R
\docType{data}
\name{hud_spec_2022}
\alias{hud_spec_2022}
\title{The most recent HUD Specifications PDF}
\format{
An object of class \code{character} of length 1.
}
\usage{
hud_spec_2022
}
\description{
The most recent HUD Specifications PDF
}
\keyword{datasets}
