% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_export_spec_extract.R
\name{hud_spec_r_type}
\alias{hud_spec_r_type}
\title{Translate HUD Export Specification Tables to useful R equivalents}
\usage{
hud_spec_r_type(hud_spec, outtype = c("chr", "hud", "fun", "typ")[1])
}
\arguments{
\item{hud_spec}{\code{(data.frame)} HUD Specification table. Output from \code{hud_export_specs}.}

\item{outtype}{\code{(character)} One of:
\itemize{
\item{\code{"chr"}}{ Returns the class as a readr abbreviation (See \link[readr]{cols})}
\item{\code{"hud"}}{ \code{(character)} a type specification from HUD}
\item{\code{"fun"}}{a readr \verb{parse_*} function (See \link[readr]{parse_logical})}{ \code{(function)}}
\item{\code{"type"}}{ \code{(character)} The R data class}
}}
}
\description{
Given various inputs, provide a col_type specification in the format indicated by \code{outtype}
}
\examples{
purrr::map(hud_export_specs(hud_pdf_data()), hud_spec_r_type)
}
