% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geocodes.R
\name{ohio_city2county}
\alias{ohio_city2county}
\title{Create a table that maps Ohio Cities to Counties}
\usage{
ohio_city2county(
  file = "https://www.omlohio.org/DocumentCenter/View/117/Municipalities-in-Ohio-PDF",
  path = file.path("data", "ohio_city2county.feather"),
  overwrite = FALSE
)
}
\arguments{
\item{file}{\code{(character)} the file path or URL. If none is specified, the \emph{inst} directory is searched for the most recent file.}

\item{path}{\code{(character)} The path to save the feather file for the extracted data.}

\item{overwrite}{\code{(logical)} Whether to overwrite the existing feather file or just read it and return it.}
}
\value{
\code{(tbl_df)} of geocodes. A message indicating where the feather file has been saved for quick retrieval.
}
\description{
This uses online references to create a table that maps Ohio Cities to Counties.
}
\details{
Extracting tables from every page of a PDF is an expensive process, therefore this function saves a  \link[feather]{feather} file for rapid retrieval. The file will be written to \code{path} with the same name as the pdf passed to \code{file}.
}
