% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geocodes.R
\name{ohio_geocodes_full}
\alias{ohio_geocodes_full}
\title{Extract HUD Geocodes for the state of Ohio}
\usage{
ohio_geocodes_full(files = historical_geocode_pdfs, overwrite = FALSE)
}
\value{
\code{(tbl_df)} of geocodes. A message indicating where the feather file has been saved for quick retrieval.
}
\description{
Extract HUD Geocodes from latest specification pdf in inst folder
}
\details{
Extracting tables from every page of a PDF is an expensive process, therefore this function saves a  \link[feather]{feather} file for rapid retrieval. The file will be written to \code{path} with the same name as the pdf passed to \code{file}.
}
