% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_open.R
\name{open_pdfs}
\alias{open_pdfs}
\title{Open stored HUD PDFs}
\usage{
open_pdfs(
  type = c("csv_specs", "data_dictionarys", "data_standards", "acronyms")[1],
  year = "latest"
)
}
\arguments{
\item{type}{\code{(character)} One of: \code{csv_specs, data_dictionarys, data_standards, acronyms}}

\item{year}{\code{(character/numeric)} to open. Use \code{"latest"} to return the most recent}
}
\value{

}
\description{
Open stored HUD PDFs
}
\examples{
open_pdfs()
}
