% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_geocodes.R
\name{hud_geocodes}
\alias{hud_geocodes}
\title{Extract HUD Geocodes}
\usage{
hud_geocodes(file)
}
\arguments{
\item{file}{\code{(character)} the file path or URL. If none is specified, the \emph{inst} directory is searched for the most recent file.}
}
\value{
\code{(tbl_df)} of geocodes
}
\description{
Extract HUD Geocodes from latest specification pdf in inst folder
}
