% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hud_value_extract.R
\name{hud_value_tables}
\alias{hud_value_tables}
\title{Create hash table functions that transform code specification integers to human-legible characters}
\usage{
hud_value_tables(
  hud_pdf_data,
  dims,
  .write = TRUE,
  verify = interactive(),
  overwrite = FALSE,
  path = file.path("data", "public", "export_text_translations")
)
}
\arguments{
\item{dims}{\code{(list)} output from \code{hud_dimensions}}

\item{\code{(list)}}{output from \code{hud_pdf_data}}
}
\description{
This function parses the appendices of the HUD Spec PDF where Value => Text correspondence tables are represented and transforms these tables into functions that allow coercion between Value/Text
}
